raw2userData.js replaced by util.js which also contains converse
